Fix note event types and add IntelliJ project files
Corrected the types for timeSeconds and seq in ChartToEngine.kt to use Double and Long, respectively. Added IntelliJ IDEA project configuration files to version control. Also included a demo chart JSON and updated build artifacts.

+++ core/src/main/kotlin/io/github/therealsegfault/projectbeatsgdx/core/chart/ChartToEngine.kt
@@ -14,9 +14,9 @@ object ChartToEngine {
 
         return sorted.mapIndexed { idx, n ->
             NoteEvent(
-                timeSeconds = n.timeSeconds.toLong(),
+                timeSeconds = n.timeSeconds.toDouble(),
                 lane = n.lane,
-                seq = idx
+                seq = idx.toLong()
             )
         }
     }
